requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchaudio
torchvision

absl-py==2.1.0
affine==2.4.0
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.13.1
fonttools==4.53.1
fsspec==2024.2.0
geopandas==1.0.1
grpcio==1.66.1
hydra-core==1.3.2
idna==3.10
imageio==2.35.1
importlib_metadata==8.4.0
importlib_resources==6.4.4
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
lazy_loader==0.4
loguru==0.7.2
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
omegaconf==2.3.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.28.1
psutil==6.0.0
pure_eval==0.2.3
Pygments==2.18.0
pyogrio==0.9.0
pyparsing==3.1.4
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
rasterio==1.3.11
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
snuggs==1.4.7
soupsieve==2.6
stack-data==0.6.3
sympy==1.12
tensorboard==2.17.1
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.8.30
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
win32-setctime==1.1.0
yarg==0.1.9
zipp==3.20.1
